Cherry-pick 310165@main (5eb7a8756f8c). https://bugs.webkit.org/show_bug.cgi?id=310998
    Factor Overflow Handlers out of CheckedArithmetic.h, and update OverflowPolicy accordingly.
    https://bugs.webkit.org/show_bug.cgi?id=310998
    radar://173602523

    Reviewed by Keith Miller.

    This makes OverflowPolicy consistent with the available Overflow Handlers (basically, adds
    AssertNoOverflow), and allows Overflow Handlers to be used by clients without having to
    include all of CheckedArithmetic.h.

    No new tests because there is no behavior change.

    * Source/WTF/WTF.xcodeproj/project.pbxproj:
    * Source/WTF/wtf/CMakeLists.txt:
    * Source/WTF/wtf/CheckedArithmetic.h:
    (WTF::AssertNoOverflow::overflowed): Deleted.
    (WTF::AssertNoOverflow::clearOverflow): Deleted.
    (WTF::AssertNoOverflow::crash): Deleted.
    (WTF::AssertNoOverflow::hasOverflowed const): Deleted.
    (WTF::CrashOnOverflow::overflowed): Deleted.
    (WTF::CrashOnOverflow::clearOverflow): Deleted.
    (WTF::CrashOnOverflow::crash): Deleted.
    (WTF::CrashOnOverflow::hasOverflowed const): Deleted.
    (WTF::RecordOverflow::RecordOverflow): Deleted.
    (WTF::RecordOverflow::clearOverflow): Deleted.
    (WTF::RecordOverflow::crash): Deleted.
    (WTF::RecordOverflow::hasOverflowed const): Deleted.
    (WTF::RecordOverflow::overflowed): Deleted.
    * Source/WTF/wtf/OverflowHandler.h: Added.
    (WTF::AssertNoOverflow::overflowed):
    (WTF::AssertNoOverflow::clearOverflow):
    (WTF::AssertNoOverflow::crash):
    (WTF::AssertNoOverflow::hasOverflowed const):
    (WTF::CrashOnOverflow::overflowed):
    (WTF::CrashOnOverflow::clearOverflow):
    (WTF::CrashOnOverflow::crash):
    (WTF::CrashOnOverflow::hasOverflowed const):
    (WTF::RecordOverflow::RecordOverflow):
    (WTF::RecordOverflow::clearOverflow):
    (WTF::RecordOverflow::crash):
    (WTF::RecordOverflow::hasOverflowed const):
    (WTF::RecordOverflow::overflowed):
    * Source/WTF/wtf/OverflowPolicy.h:
    (WTF::shouldCrashOnOverflow):
    * Source/WTF/wtf/text/StringBuilder.h:
    (WTF::StringBuilder::StringBuilder):

    Canonical link: https://commits.webkit.org/310165@main
