+++ Source/WTF/wtf/CMakeLists.txt
@@ -216,6 +216,7 @@ set(WTF_PUBLIC_HEADERS
     OptionalOrReference.h
     OptionSetHash.h
     OrderMaker.h
+    OverflowHandler.h
     OverflowPolicy.h
     Packed.h
     PackedRef.h
